Teacher and Student (address, validation)

+++ src/main/java/com/school/pfe/Service/impl/TeacherServiceImpl.java
@@ -2,7 +2,17 @@ package com.school.pfe.Service.impl;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import com.school.pfe.Dto.SubjectDto;
+import com.school.pfe.Exception.EntityNotFoundException;
+import com.school.pfe.Exception.ErrorCodes;
+import com.school.pfe.Exception.InvalidEntityException;
+import com.school.pfe.Exception.InvalidOperationException;
+import com.school.pfe.Model.Subject;
+import com.school.pfe.Validation.SubjectValidation;
+import com.school.pfe.Validation.TeacherValidation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +21,7 @@ import com.school.pfe.Model.Teacher;
 import com.school.pfe.Repository.TeacherRepository;
 import com.school.pfe.Service.TeacherService;
 @Service
+@Slf4j
 public class TeacherServiceImpl implements TeacherService{
 
 	@Autowired
@@ -19,33 +30,48 @@ public class TeacherServiceImpl implements TeacherService{
 	
 	@Override
 	public TeacherDto save(TeacherDto teacherDto) {
-		Teacher teacher=TeacherDto.dtoToEntity(teacherDto);
-		teacherRepository.save(teacher);
-		TeacherDto teacherDtosaved=TeacherDto.entityToDto(teacher);
-		return teacherDtosaved;
+		List<String> errors = TeacherValidation.validate(teacherDto);
+
+		if (!errors.isEmpty()) {
+			log.error("teacher is not valid {}", teacherDto);
+			throw new InvalidEntityException("Le professeur n'est pas valide", ErrorCodes.TEACHER_NOT_VALID, errors);
+		}
+		return TeacherDto.entityToDto(
+				teacherRepository.save(
+						TeacherDto.dtoToEntity(teacherDto)
+				)
+		);
 	}
 
 	@Override
 	public TeacherDto teacherById(Long id) {
-		Optional<Teacher> optionalTeacher =teacherRepository.findById(id);
-		 if (optionalTeacher.isPresent()) {
-			 Teacher teacher=optionalTeacher.get();
-				 return TeacherDto.entityToDto(teacher);
-		 }
-		 else
-			{
-				return null;
-			}
+		if (id == null) {
+			log.error("Teacher ID is null");
+			throw new InvalidOperationException("Impossible de trouver un professeur avec un ID NULL", ErrorCodes.TEACHER_ID_IS_NULL);
+		}
+		Optional<Teacher> teacher = teacherRepository.findById(id);
+
+		return Optional.of(TeacherDto.entityToDto(teacher.get())).orElseThrow(() ->
+				new EntityNotFoundException(
+						"Aucun professeur avec l'ID = " + id + " n'a été trouvé dans la BDD",
+						ErrorCodes.TEACHER_NOT_FOUND)
+		);
 		 }
 
 	@Override
 	public List<TeacherDto> findAll() {
-		return TeacherDto.entityToDto(teacherRepository.findAll());
+		return teacherRepository.findAll().stream()
+				.map(TeacherDto::entityToDto)
+				.collect(Collectors.toList());
 	}
 
 	@Override
 	public void delete(Long id) {
-		teacherRepository.deleteById(id);		
+		if (id == null) {
+			log.error("Teacher ID is null");
+			throw new InvalidOperationException("Impossible de supprimer un professeur  avec un ID NULL",ErrorCodes.TEACHER_ID_IS_NULL);
+		}
+		teacherRepository.deleteById(id);
 	}
 
 }
